backend/src/requirements.txt
asgiref==3.8.1
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.8
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.1
PyYAML==6.0.2
requests==2.32.3
setuptools==74.1.2
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
